+++ examples/counter/src/androidMain/kotlin/com/copperleaf/ballast/examples/counter/platformActual.android.kt
@@ -1,8 +1,11 @@
 package com.copperleaf.ballast.examples.counter
 
 import com.copperleaf.ballast.BallastLogger
+import com.copperleaf.ballast.BallastViewModelConfiguration
 import com.copperleaf.ballast.core.AndroidLogger
 import com.copperleaf.ballast.debugger.BallastDebuggerClientConnection
+import com.copperleaf.ballast.debugger.BallastDebuggerInterceptor
+import com.copperleaf.ballast.plusAssign
 import io.ktor.client.engine.cio.CIO
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -18,9 +21,10 @@ private val lazyConnection by lazy {
     }.also { it.connect() }
 }
 
-internal actual fun platformDebuggerConnection(): BallastDebuggerClientConnection<*> {
-    return lazyConnection
-}
+internal actual fun BallastViewModelConfiguration.Builder.installDebugger(): BallastViewModelConfiguration.Builder =
+    apply {
+        this += BallastDebuggerInterceptor(lazyConnection)
+    }
 
 internal actual fun platformLogger(loggerName: String): BallastLogger {
     return AndroidLogger(loggerName)